backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
python-whois==0.9.4
requests==2.31.0
python-dotenv==1.0.1
supabase==2.4.0
redis==5.0.4
httpx==0.27.0
dnspython==2.6.1
